# docs/api-specification.yml
openapi: 3.0.3
info:
  title: テストケース管理システム API
  description: |
    Next.js (App Router) + PostgreSQL + NextAuth.jsを使用したテストケース管理システムのREST API仕様書

    ## 認証
    すべてのAPIエンドポイント（認証関連を除く）はJWT認証が必要です。
    NextAuth.jsのセッショントークンを使用してリクエストを行います。

    ## 権限管理
    ### 静的ロール (user_role)
    - 0: 管理者 - すべての操作が可能
    - 1: テスト管理者 - テストグループの作成が可能
    - 2: 一般 - 割り当てられたテストグループのみアクセス可能

    ### 動的ロール (test_role)
    - 1: 設計者 - テストケースの作成・編集が可能
    - 2: 実施者 - テスト結果の登録が可能
    - 3: 閲覧者 - 閲覧のみ

  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:3000/api
    description: 開発環境
  - url: https://api.example.com
    description: 本番環境
...
paths:
...
  /test-groups/{groupId}/cases:
    get:
      tags:
        - Test Cases
      summary: テストケース一覧取得
      description: |
        指定したテストグループのテストケース一覧を取得
        **実装状態**: ✅ 完了
      operationId: getTestCases
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/GroupId'
      responses:
        '200':
          description: テストケース一覧
          content:
            application/json:
              schema:
                type: object
                properties:
                  testCases:
                    type: array
                    items:
                      $ref: '#/components/schemas/TestCaseWithContents'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags:
        - Test Cases
      summary: テストケース作成
      description: |
        新しいテストケースを作成（設計者ロールまたは管理者のみ）
        **実装状態**: ✅ 完了
      operationId: createTestCase
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/GroupId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestCaseInput'
      responses:
        '201':
          description: テストケース作成成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  testCase:
                    $ref: '#/components/schemas/TestCase'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /test-groups/{groupId}/cases/{tid}:
    get:
      tags:
        - Test Cases
      summary: テストケース詳細取得
      description: |
        指定したテストケースの詳細情報を取得
        **実装状態**: 🔨 未実装（要実装）
      operationId: getTestCase
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/GroupId'
        - $ref: '#/components/parameters/Tid'
      responses:
        '200':
          description: テストケース詳細
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestCaseWithContents'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Test Cases
      summary: テストケース更新
      description: |
        テストケース情報を更新（設計者ロールまたは管理者のみ）
        **実装状態**: 🔨 未実装（要実装）
      operationId: updateTestCase
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/GroupId'
        - $ref: '#/components/parameters/Tid'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestCaseInput'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  testCase:
                    $ref: '#/components/schemas/TestCase'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Test Cases
      summary: テストケース削除
      description: |
        テストケースをソフトデリート
        **実装状態**: 🔨 未実装（要実装）
      operationId: deleteTestCase
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/GroupId'
        - $ref: '#/components/parameters/Tid'
      responses:
        '200':
          description: 削除成功
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==========================================
  # Test Results
  # ==========================================
  /test-groups/{groupId}/cases/{tid}/results:
    get:
      tags:
        - Test Results
      summary: テスト結果一覧取得
      description: |
        指定したテストケースのテスト結果一覧を取得
        **実装状態**: 🔨 未実装（要実装）
      operationId: getTestResults
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/GroupId'
        - $ref: '#/components/parameters/Tid'
      responses:
        '200':
          description: テスト結果一覧
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/TestResultWithContent'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: NextAuth.js JWT トークン
  parameters:
    GroupId:
      name: groupId
      in: path
      required: true
      description: テストグループID
      schema:
        type: integer
        example: 1
    Tid:
      name: tid
      in: path
      required: true
      description: テストケースID (TID)
      schema:
        type: string
        example: TC-001
...
  schemas:
...
    TestCase:
      type: object
      properties:
        test_group_id:
          type: integer
        tid:
          type: string
        first_layer:
          type: string
        second_layer:
          type: string
        third_layer:
          type: string
        fourth_layer:
          type: string
        purpose:
          type: string
        request_id:
          type: string
        check_items:
          type: string
        test_procedure:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    TestCaseWithContents:
      allOf:
        - $ref: '#/components/schemas/TestCase'
        - type: object
          properties:
            contents:
              type: array
              items:
                $ref: '#/components/schemas/TestContent'
            files:
              type: array
              items:
                $ref: '#/components/schemas/TestCaseFile'
    TestCaseInput:
      type: object
      required:
        - tid
      properties:
        tid:
          type: string
        first_layer:
          type: string
        second_layer:
          type: string
        third_layer:
          type: string
        fourth_layer:
          type: string
        purpose:
          type: string
        request_id:
          type: string
        check_items:
          type: string
        test_procedure:
          type: string
        contents:
          type: array
          items:
            type: object
            properties:
              test_case_no:
                type: integer
              test_case:
                type: string
              expected_value:
                type: string
              is_target:
                type: boolean
        files:
          type: array
          items:
            type: object
            properties:
              file_type:
                type: integer
                description: 0=制御仕様, 1=データフロー
              file_no:
                type: integer
              file_name:
                type: string
              file_path:
                type: string
                description: S3パス
    TestContent:
      type: object
      properties:
        test_case_no:
          type: integer
        test_case:
          type: string
        expected_value:
          type: string
        is_target:
          type: boolean
    TestCaseFile:
      type: object
      properties:
        file_type:
          type: integer
        file_no:
          type: integer
        file_name:
          type: string
        file_path:
          type: string

    # Test Results
    TestResult:
      type: object
      properties:
        test_group_id:
          type: integer
        tid:
          type: string
        test_case_no:
          type: integer
        result:
          type: string
        judgment:
          type: string
          enum: [OK, NG, 再実施対象外]
        software_version:
          type: string
        hardware_version:
          type: string
        comparator_version:
          type: string
        execution_date:
          type: string
          format: date
        executor:
          type: string
        note:
          type: string
        version:
          type: integer
...
    TestResultWithContent:
      allOf:
        - $ref: '#/components/schemas/TestResult'
        - type: object
          properties:
            test_case:
              type: string
            expected_value:
              type: string
            is_target:
              type: boolean
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: エラーメッセージ
  responses:
    BadRequest:
      description: リクエストが不正です
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: 認証が必要です
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: 権限がありません
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: リソースが見つかりません
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
security:
  - bearerAuth: []